gherkin: (C) Conform to C90 (mostly)
To support a wider range of compilers, remove the C90 breaches so that
the code compiles also with -ansi and/or -std=c90 (using gcc).
Another reason to not use swprintf is that on many Windows compilers
swprintf has a different signature than the ISO standard specification.
When cleaning up commented out code in gherkin_generate_tokens.c (to
remove the usage of // for comments), settle to use the
FileTokenScanner to get coverage of it in the acceptance tests.

The code does not compile with -std=c90 -pedantic (using gcc), the
conformance to C90 is not taken that far.